@@ -24,23 +24,37 @@
       "last 1 safari version"
     ]
   },
+  "overrides": {
+    "oidc-token-hash": "5.0.1"
+  },
   "dependencies": {
     "@emailjs/browser": "^3.11.0",
     "@ffmpeg/ffmpeg": "^0.12.6",
     "@ffmpeg/util": "^0.12.1",
     "@headlessui/react": "^1.7.17",
+    "@metaplex-foundation/digital-asset-standard-api": "^1.0.0",
     "@metaplex-foundation/js": "^0.19.4",
+    "@metaplex-foundation/mpl-bubblegum": "^3.1.2",
+    "@metaplex-foundation/mpl-toolbox": "^0.9.2",
+    "@metaplex-foundation/umi-bundle-defaults": "^0.9.1",
+    "@metaplex-foundation/umi-signer-wallet-adapters": "^0.9.1",
     "@metaplex/js": "^4.12.0",
     "@mux/mux-node": "^7.3.2",
     "@mux/mux-player": "^2.0.1",
     "@mux/mux-player-react": "^2.0.1",
+    "@project-serum/anchor": "^0.26.0",
     "@react-three/cannon": "^6.5.2",
     "@react-three/drei": "^9.65.4",
     "@react-three/fiber": "^8.12.1",
     "@redux-devtools/extension": "^3.2.5",
+    "@solana/spl-account-compression": "^0.2.0",
+    "@solana/spl-token": "^0.4.0",
+    "@solana/wallet-adapter-base": "^0.9.23",
     "@solana/wallet-adapter-nextjs-starter": "^0.8.17",
-    "@solana/web3.js": "^1.78.4",
-    "@spling/social-protocol": "^1.5.9",
+    "@solana/wallet-adapter-react": "^0.15.35",
+    "@solana/wallet-adapter-react-ui": "^0.9.33",
+    "@solana/wallet-adapter-wallets": "^0.19.25",
+    "@solana/web3.js": "^1.89.1",
     "@types/react": "^18.0.32",
     "@types/react-dom": "18.0.11",
     "@types/three": "^0.155.0",
@@ -48,26 +62,28 @@
     "@typescript-eslint/parser": "^6.5.0",
     "@underdog-protocol/passport": "^4.0.0",
     "@vercel/kv": "^1.0.0",
+    "@wavesurfer/react": "^1.0.4",
     "arweave": "^1.13.7",
     "auth": "^0.0.9",
     "bs58": "^5.0.0",
     "dotenv": "^16.3.1",
     "eslint": "^8.37.0",
-    "eslint-config-next": "13.2.4",
-    "fluent-ffmpeg": "^2.1.2",
     "formidable": "^3.5.1",
     "framer-motion": "^10.12.4",
     "fs": "^0.0.1-security",
+    "lottie-web": "^5.12.2",
     "maplibre-gl": "^3.5.1",
     "micro": "^10.0.1",
     "mongodb": "^5.3.0",
-    "next": "^12.3.4",
+    "next": "^14.0.4",
     "next-auth": "^4.24.5",
+    "node-fetch": "^3.3.2",
     "party-js": "^2.2.0",
     "passport-google-oidc": "^0.1.0",
+    "pino-pretty": "^10.2.3",
     "querystring": "^0.2.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hot-toast": "^2.4.1",
     "react-map-gl": "^7.1.6",
     "react-use": "^17.4.0",
@@ -75,7 +91,7 @@
     "sass": "^1.65.1",
     "sonner": "^1.2.3",
     "three": "^0.151.3",
-    "wavesurfer.js": "^7.4.2",
+    "wavesurfer.js": "^7.7.1",
     "zustand": "^4.4.4"
   },
   "devDependencies": {
@@ -83,6 +99,7 @@
     "autoprefixer": "^10.4.14",
     "babel-loader": "^9.1.3",
     "daisyui": "^3.6.3",
+    "eslint-config-next": "^14.0.4",
     "eslint-config-prettier": "^8.8.0",
     "postcss": "^8.4.21",
     "prettier": "^2.8.7",
