@@ -1,6 +1,6 @@
 import React from "react";
 import Map, { Marker } from "react-map-gl";
-import Image from "next/image";
+import Image from "next/legacy/image";
 
 interface MapViewProps {
   latitude?: number;
@@ -16,18 +16,21 @@ export const MapView: React.FC<MapViewProps> = ({
   shouldZoom,
 }) => {
   return (
-    <div className="h-full flex justify-center items-center mx-4 mb-4 rounded-md ">
+    <div
+      className="flex justify-center items-center w-full rounded-md"
+      style={{ height: "calc(100dvh - 8rem)" }}
+    >
       <Map
         mapboxAccessToken={process.env.NEXT_PUBLIC_MAPBOX_API_KEY}
         mapStyle={`mapbox://styles/noamie22/ck2zynqqn018x1cnohcw5cnna`}
         initialViewState={{
           longitude: Number(longitude),
           latitude: Number(latitude),
-          zoom: 10,
+          zoom: shouldZoom ? 20 : 15,
         }}
       >
         {!markers && longitude && latitude ? (
-          <Marker longitude={longitude} latitude={latitude} color="red">
+          <Marker longitude={longitude} latitude={latitude}>
             <Image
               src={"/ear_small.png"}
               alt="øhr logo"
